[MIG] account_bank_statement_import_online_stripe: Migration to 13.0
