+++ account_bank_statement_import_online_stripe/README.rst
@@ -1,38 +1,32 @@
-.. image:: https://odoo-community.org/readme-banner-image
-   :target: https://odoo-community.org/get-involved?utm_source=readme
-   :alt: Odoo Community Association
-
 ==============================
 Online Bank Statements: Stripe
 ==============================
 
-.. 
+..
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:cf1b7d039ce6fc8161aa033040c67fec12c9fec42818ce5dd19ccbc89d681797
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
     :target: https://odoo-community.org/page/development-status
     :alt: Beta
-.. |badge2| image:: https://img.shields.io/badge/license-AGPL--3-blue.png
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fbank--statement--import-lightgray.png?logo=github
-    :target: https://github.com/OCA/bank-statement-import/tree/18.0/account_statement_import_online_stripe
+    :target: https://github.com/OCA/bank-statement-import/tree/13.0/account_bank_statement_import_online_stripe
     :alt: OCA/bank-statement-import
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/bank-statement-import-18-0/bank-statement-import-18-0-account_statement_import_online_stripe
+    :target: https://translation.odoo-community.org/projects/bank-statement-import-13-0/bank-statement-import-13-0-account_bank_statement_import_online_stripe
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
-    :target: https://runboat.odoo-community.org/builds?repo=OCA/bank-statement-import&target_branch=18.0
+    :target: https://runboat.odoo-community.org/builds?repo=OCA/bank-statement-import&target_branch=13.0
     :alt: Try me on Runboat
 
 |badge1| |badge2| |badge3| |badge4| |badge5|
 
-This module provides online bank statements from stripe.com
+This module provides online bank statements from Stripe.
 
 **Table of contents**
 
@@ -42,16 +36,14 @@ This module provides online bank statements from stripe.com
 Configuration
 =============
 
-1. Ensure your user has the necessary permissions. Assign the user to
-   the group "Technical / Show Full Accounting Features" to gain access
-   to the required settings.
-2. Go to "Invoicing" -> "Configuration" -> "Online Bank Statement
-   Providers".
-3. Configure the Provider
+1. Ensure your user has the necessary permissions. Assign the user to the group
+   "Technical / Show Full Accounting Features" to gain access to the required settings.
+2. Go to *Invoicing* -> *Configuration* -> *Online Bank Statement Providers*.
+3. Configure the Provider.
 
-For more detailed information about the Stripe API and the balance
-transactions used by this module, please refer to the official Stripe
-documentation: https://docs.stripe.com/api/balance_transactions
+For more detailed information about the Stripe API and the balance transactions used by
+this module, please refer to the official Stripe documentation:
+https://docs.stripe.com/api/balance_transactions
 
 Bug Tracker
 ===========
@@ -59,7 +51,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/bank-statement-import/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/bank-statement-import/issues/new?body=module:%20account_statement_import_online_stripe%0Aversion:%2018.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/bank-statement-import/issues/new?body=module:%20account_bank_statement_import_online_stripe%0Aversion:%2013.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -67,24 +59,24 @@ Credits
 =======
 
 Authors
--------
+~~~~~~~
 
 * Akiles
 * Tecnativa
 
 Contributors
-------------
+~~~~~~~~~~~~
 
-- `Akiles <https://akiles.app/>`__:
+* `Akiles <https://akiles.app/>`__:
 
-  - Dario Nieuwenhuis
+  * Dario Nieuwenhuis
 
-- `Tecnativa <https://www.tecnativa.com>`__:
+* `Tecnativa <https://www.tecnativa.com>`__:
 
-  - Juan Carlos Oñate
+  * Juan Carlos Oñate
 
 Maintainers
------------
+~~~~~~~~~~~
 
 This module is maintained by the OCA.
 
@@ -102,8 +94,8 @@ promote its widespread use.
 
 Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
 
-|maintainer-juancarlosonate-tecnativa| 
+|maintainer-juancarlosonate-tecnativa|
 
-This module is part of the `OCA/bank-statement-import <https://github.com/OCA/bank-statement-import/tree/18.0/account_statement_import_online_stripe>`_ project on GitHub.
+This module is part of the `OCA/bank-statement-import <https://github.com/OCA/bank-statement-import/tree/13.0/account_bank_statement_import_online_stripe>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
