+++ account_bank_statement_import_online_stripe/models/online_bank_statement_provider_stripe.py
@@ -9,7 +9,7 @@ from urllib.parse import urlencode
 
 import pytz
 
-from odoo import api, fields, models
+from odoo import _, api, fields, models
 from odoo.exceptions import UserError
 
 STRIPE_API_BASE = "https://api.stripe.com/v1"
@@ -70,30 +70,27 @@ class OnlineBankStatementProviderStripe(models.Model):
                 continue
             lines.append(
                 {
+                    "name": safe_format(self.stripe_label, tx),
+                    "note": safe_format(self.stripe_note, tx),
                     "ref": safe_format(self.stripe_reference, tx),
-                    "payment_ref": safe_format(self.stripe_label, tx),
-                    "narration": safe_format(self.stripe_note, tx),
                     "amount": float(tx["amount"]) / (10 ** currency.decimal_places),
                     "date": datetime.fromtimestamp(tx["created"]),
                     "unique_import_id": tx["id"],
-                    "raw_data": json.dumps(tx),
                 }
             )
             if tx.get("fee"):
                 lines.append(
                     {
+                        "name": safe_format(self.stripe_fee_label, tx),
+                        "note": safe_format(self.stripe_fee_note, tx),
                         "ref": safe_format(self.stripe_fee_reference, tx),
-                        "payment_ref": safe_format(self.stripe_fee_label, tx),
-                        "narration": safe_format(self.stripe_fee_note, tx),
                         "amount": float(-tx["fee"]) / (10 ** currency.decimal_places),
                         "date": datetime.fromtimestamp(tx["created"]),
                         "unique_import_id": tx["id"] + "_fee",
-                        "raw_data": json.dumps(tx),
                     }
                 )
         return lines, {}
 
-    @api.model
     def _stripe_api_get_all(self, path, params=None):
         if params is None:
             params = []
@@ -109,7 +106,6 @@ class OnlineBankStatementProviderStripe(models.Model):
             else:
                 break
 
-    @api.model
     def _stripe_api_get(self, path, params=None, data=None):
         if params is None:
             params = {}
@@ -118,19 +114,21 @@ class OnlineBankStatementProviderStripe(models.Model):
                 (self.api_base or STRIPE_API_BASE) + path + "?" + urlencode(params),
                 data=data,
             )
-            request.add_header("Authorization", f"Bearer {self.password}")
+            request.add_header("Authorization", "Bearer %s" % self.password)
             response = urllib.request.urlopen(request)
             content = response.read().decode("utf-8")
             return json.loads(content)
         except HTTPError as e:
             content = json.loads(e.read().decode("utf-8"))
-            raise UserError(f"Stripe API call failed: {path}: {content}") from e
+            raise UserError(
+                _("Stripe API call failed: %s: %s") % (path, content)
+            ) from e
 
 
 def safe_format(template, kwargs):
     template = template or ""
 
-    def sub(m: re.Match):
+    def sub(m):
         val = kwargs
         for k in m.group(1).split("."):
             try:
