--- account_statement_import_online_stripe/i18n/account_statement_import_online_stripe.pot
@@ -1,104 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* account_statement_import_online_stripe
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 18.0\n"
-"Report-Msgid-Bugs-To: \n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "API Key"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "API base"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Expand"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model,name:account_statement_import_online_stripe.model_online_bank_statement_provider
-msgid "Online Bank Statement Provider"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Statement Fee Label"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Statement Fee Note"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Statement Fee Reference"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Statement Label"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Statement Note"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Statement Reference"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_expand
-msgid "Stripe Expand"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_fee_label
-msgid "Stripe Fee Label"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_fee_note
-msgid "Stripe Fee Note"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_fee_reference
-msgid "Stripe Fee Reference"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_label
-msgid "Stripe Label"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_note
-msgid "Stripe Note"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_reference
-msgid "Stripe Reference"
-msgstr ""
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Stripe configuration"
-msgstr ""
