--- account_statement_import_online_stripe/i18n/it.po
@@ -1,107 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* account_statement_import_online_stripe
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 17.0\n"
-"Report-Msgid-Bugs-To: \n"
-"PO-Revision-Date: 2025-04-08 09:24+0000\n"
-"Last-Translator: mymage <stefano.consolaro@mymage.it>\n"
-"Language-Team: none\n"
-"Language: it\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=2; plural=n != 1;\n"
-"X-Generator: Weblate 5.10.4\n"
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "API Key"
-msgstr "Chiave API"
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "API base"
-msgstr "API base"
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Expand"
-msgstr "Espandi"
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model,name:account_statement_import_online_stripe.model_online_bank_statement_provider
-msgid "Online Bank Statement Provider"
-msgstr "Fornitore estratto conto bancario online"
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Statement Fee Label"
-msgstr "Etichetta tariffa estratto conto"
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Statement Fee Note"
-msgstr "Nota tariffa estratto conto"
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Statement Fee Reference"
-msgstr "Riferimento tariffa estratto conto"
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Statement Label"
-msgstr "Etichetta estratto conto"
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Statement Note"
-msgstr "Nota estratto conto"
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Statement Reference"
-msgstr "Riferimento estratto conto"
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_expand
-msgid "Stripe Expand"
-msgstr "Espandi transizione"
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_fee_label
-msgid "Stripe Fee Label"
-msgstr "Etichetta tariffa transizione"
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_fee_note
-msgid "Stripe Fee Note"
-msgstr "Nota tariffa transizione"
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_fee_reference
-msgid "Stripe Fee Reference"
-msgstr "Riferimento tariffa transizione"
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_label
-msgid "Stripe Label"
-msgstr "Etichetta transizione"
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_note
-msgid "Stripe Note"
-msgstr "Nota transizione"
-
-#. module: account_statement_import_online_stripe
-#: model:ir.model.fields,field_description:account_statement_import_online_stripe.field_online_bank_statement_provider__stripe_reference
-msgid "Stripe Reference"
-msgstr "Riferimento transizione"
-
-#. module: account_statement_import_online_stripe
-#: model_terms:ir.ui.view,arch_db:account_statement_import_online_stripe.online_bank_statement_provider_form
-msgid "Stripe configuration"
-msgstr "Configurazione transizione"
