+++ setup/account_bank_statement_import_online_stripe/odoo/addons/account_bank_statement_import_online_stripe
@@ -1,0 +1,1 @@
+../../../../account_bank_statement_import_online_stripe
