--- setup/account_statement_import_online_stripe/odoo/addons/account_statement_import_online_stripe
@@ -1,1 +1,0 @@
-../../../../account_statement_import_online_stripe
